+++ Assets/Scenes/NPC 개발씬.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7396c9c3d34017247b1ff77c2170b4f3
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
