+++ Assets/Scenes/맵 개발 씬.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: eff21be78e912f74dae4dd33bf9de284
-NativeFormatImporter:
+guid: 2151578b7438d124da7f9e5b3b02cb00
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
