+++ Assets/Scenes/메인메뉴 개발 씬.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 4e6b71cdaefdd944593df0615e644af7
-NativeFormatImporter:
+guid: a6be48851a86a4b4e9c1e00f796a43c5
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
