+++ Assets/Units/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3c7903269bbab34884f8c0069011040
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
