+++ Assets/Units/Player.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 5045bdde4d7925f43bc0b460d9ac6fc1
+guid: 16c5dd53fec3fc9499a291e9e20be46c
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
