+++ Assets/Units/맵/블록-잔디.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 695ba11c562c245429e4b0d532543fa5
-folderAsset: yes
-DefaultImporter:
+guid: 3df669eece153444face589b9a4b1baa
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
