+++ Assets/Units/플레이어.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f784bdbc7471cf349bb7a2f29eac55f0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
