+++ Assets/Units/플레이어/Hero_Knight/Sprites/anim/Idle_0.controller.meta
