+++ Assets/Units/플레이어/플레이어.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 9fb5e3eacd850534bbdfd9070170b110
-folderAsset: yes
-DefaultImporter:
+guid: 080aa96270646494880a200bf21522e8
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
